modify float type in graph lib to double type

+++ src/lib/graph.c
@@ -55,7 +55,7 @@ void print_graph(graph *g, int simple) {
 
 /******************************* node *******************************/
 
-node *create_node(float x, float y, float radius, int r, int g, int b, void *extra) {
+node *create_node(double x, double y, double radius, int r, int g, int b, void *extra) {
     node *n = malloc(sizeof(node));
     n->x = x;
     n->y = y;
@@ -74,15 +74,15 @@ int set_node_color(node *n, int r, int g, int b) {
     return 0;
 }
 
-float get_node_x(node *n) {
+double get_node_x(node *n) {
     return n->x;
 }
 
-float get_node_y(node *n) {
+double get_node_y(node *n) {
     return n->y;
 }
 
-float get_node_radius(node *n) {
+double get_node_radius(node *n) {
     return n->radius;
 }
 
@@ -102,17 +102,17 @@ void *get_node_extra(node *n) {
     return n->extra;
 }
 
-int set_node_x(node *n, float x) {
+int set_node_x(node *n, double x) {
     n->x = x;
     return 0;
 }
 
-int set_node_y(node *n, float y) {
+int set_node_y(node *n, double y) {
     n->y = y;
     return 0;
 }
 
-int set_node_radius(node *n, float radius) {
+int set_node_radius(node *n, double radius) {
     n->radius = radius;
     return 0;
 }
